Move login component into components folder for readability
