+++ App.js
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { View, Text } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import Login from './Login.js';
+import Login from './Components/Login.js';
 
 const Stack = createStackNavigator();
